Fixed positioning issue / Menu close
The positioning of the menu was obstructing the button, thus the menu
could not be closed. I moved the menu below the header so the menu can
now be opened and closed easily. I have alos merged the css files,
created a reset.css file and removed superfluous stylesheets.
Additional commenting included in new stylesheets.